refactor: create seperate file fro Sequoia form template admin form js
