+++ assets/src/js/admin/admin.js
@@ -23,6 +23,7 @@ import './admin-export';
 import './admin-scripts';
 import './admin-importer';
 import './shortcode-button';
+import './form-template/edit-form';
 
 // Stripe core.
 import './stripe-admin';
